// frontend/src/app/model/Filiale.ts
export class Filiale {
    
    fili_id: Number;
	fili_numero: Number;
	fili_codecompagnie: String;
    fili_codepays: Number;
	fili_codegroupe: String;
	fili_denomination: String ;
	fili_sigle: String;
	fili_codedevise: String;
	fili_adresse1: String;
	fili_adresse2: String;
	fili_telephone1: String;
	fili_telephone2: String;
	fili_telephonemobile: String;
	fili_codepostal: String;
	fili_codeutilisateur: String;
	fili_datemodification: Date;

}